+++ packages/cf-deploy-config-writer/src/constants.ts
@@ -1,6 +1,5 @@
 import { UI5_DEFAULT } from '@sap-ux/ui5-config';
 import { t } from './i18n';
-
 export const WelcomeFile = 'welcomeFile';
 export const XSAppFile = 'xs-app.json';
 export const XSSecurityFile = 'xs-security.json';
@@ -9,6 +8,8 @@ export const MTABuildParams = 'build-parameters';
 export const MTAFileExtension = 'mta-ext.mtaext';
 export const DefaultServiceURL = '${default-url}';
 export const ManagedXSUAA = 'managed:xsuaa';
+export const ManagedAppFront = 'managed:app-front';
+export const ManagedDestination = 'destination';
 export const HTML5RepoHost = 'html5-apps-repo:app-host';
 export const SRV_API = 'srv-api';
 export const DefaultMTADestination = 'fiori-default-srv-api';
@@ -21,7 +22,8 @@ export const CloudFoundry = 'cf';
 export const deployMode = 'deploy_mode';
 export const enableParallelDeployments = 'enable-parallel-deployments';
 export const CDSExecutable = 'cds';
-export const CDSPackage = '@sap/cds-dk';
+export const CDSDKPackage = '@sap/cds-dk';
+export const CDSPackage = '@sap/cds';
 export const MTAExecutable = 'mta';
 export const MTAPackage = 'mta';
 export const MTAPackageVersion = '^1.2.27';
@@ -56,6 +58,15 @@ export const UI5Destination = {
     URL: UI5_DEFAULT.SAPUI5_CDN,
     ProxyType: 'Internet',
     Authentication: 'NoAuthentication'
+};
+export const UI5AppfrontDestinationParameter = {
+    name: 'ui5',
+    url: UI5_DEFAULT.SAPUI5_CDN
+};
+export const CAPAppfrontDestination = {
+    name: ResourceMTADestination,
+    url: '~{srv-api/srv-url}',
+    forwardAuthToken: true
 };
 export const UI5ResourceDestination = {
     'init_data': {
@@ -117,5 +128,5 @@ export const rootDeployMTAScript = (args: string[]): string => {
 export const undeployMTAScript = (mtaId: string): string =>
     `cf undeploy ${mtaId} --delete-services --delete-service-keys --delete-service-brokers`;
 const cannotFindBinary = (bin: string, pkg: string): string => t('error.cannotFindBinary', { bin, pkg });
-export const CDSBinNotFound = cannotFindBinary(CDSExecutable, CDSPackage);
+export const CDSBinNotFound = cannotFindBinary(CDSExecutable, CDSDKPackage);
 export const MTABinNotFound = cannotFindBinary(MTAExecutable, MTAPackage);