+++ packages/fiori-app-sub-generator/test/int/fiori-elements/expected-output/alp_v4/webapp/test/integration/FirstJourney.js
@@ -1,0 +1,38 @@
+sap.ui.define([
+    "sap/ui/test/opaQunit"
+], function (opaTest) {
+    "use strict";
+
+    var Journey = {
+        run: function() {
+            QUnit.module("First journey");
+
+            opaTest("Start application", function (Given, When, Then) {
+                Given.iStartMyApp();
+
+                Then.onTheSalesOrderItemList.iSeeThisPage();
+
+            });
+
+
+            opaTest("Navigate to ObjectPage", function (Given, When, Then) {
+                // Note: this test will fail if the ListReport page doesn't show any data
+                
+                When.onTheSalesOrderItemList.onFilterBar().iExecuteSearch();
+                
+                Then.onTheSalesOrderItemList.onTable().iCheckRows();
+
+                When.onTheSalesOrderItemList.onTable().iPressRow(0);
+                Then.onTheSalesOrderItemObjectPage.iSeeThisPage();
+
+            });
+
+            opaTest("Teardown", function (Given, When, Then) { 
+                // Cleanup
+                Given.iTearDownMyApp();
+            });
+        }
+    }
+
+    return Journey;
+});