+++ packages/fiori-app-sub-generator/test/int/fiori-freestyle/expected-output/simple/webapp/controller/App.controller.js
@@ -1,0 +1,14 @@
+sap.ui.define(
+    [
+        "sap/ui/core/mvc/Controller"
+    ],
+    function(BaseController) {
+      "use strict";
+  
+      return BaseController.extend("simplenamespace.simple.controller.App", {
+        onInit: function() {
+        }
+      });
+    }
+  );
+