+++ packages/fiori-elements-writer/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @sap-ux/fiori-elements-writer
 
+## 2.1.38
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.10.2
+-   @sap-ux/cap-config-writer@0.9.25
+
+## 2.1.37
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.10.1
+-   @sap-ux/cap-config-writer@0.9.24
+
+## 2.1.36
+
+### Patch Changes
+
+-   Updated dependencies [23e055a]
+    -   @sap-ux/fiori-generator-shared@0.10.0
+    -   @sap-ux/cap-config-writer@0.9.23
+
 ## 2.1.35
 
 ### Patch Changes