+++ packages/preview-middleware/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @sap-ux/preview-middleware
 
+## 0.19.20
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.13.28
+
+## 0.19.19
+
+### Patch Changes
+
+-   Updated dependencies [d638daa]
+    -   @sap-ux/btp-utils@1.0.3
+    -   @sap-ux/adp-tooling@0.13.27
+    -   @sap-ux/system-access@0.5.35
+
+## 0.19.18
+
+### Patch Changes
+
+-   a64c215: feat: Change Table Actions CPE quick action added for ADP projects with OData V2 and V4
+
+## 0.19.17
+
+### Patch Changes
+
+-   Updated dependencies [c4278fb]
+    -   @sap-ux/adp-tooling@0.13.26
+
+## 0.19.16
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.13.25
+
+## 0.19.15
+
+### Patch Changes
+
+-   fccd2c2: feat: refactor types
+
 ## 0.19.14
 
 ### Patch Changes